Fog Work
Made a Room script.
All "rooms" have this attached to them and a light.
When an ally enters a room it turns on the light and activates any enemies in that room.
When the last ally exits a room it dims the room light.
All allies now have a rigidbody2d and a box collider 2d.
2d lights are now used.