Worked on swapping
Added an event that goes off when a character is selected.

The skill swap butt becomes non interactable when an enemy is selected, an ally without a skill is selected, or an ally with a skill on cooldown is selected.

I need to now make it so that cooldowns actually work