New Enemy AI
Finished reworking the enemy AI and replacing references with events in enemyAI.

There seems to be a few strange bugs that need to be worked out, but otherwise seems to be working fine.

Obviously I will have to work out these bugs, but I also need to continue changing references to events.