Move stuff around and bug fixes
Fixed a mini map bug. Put in some of the new UI sprites (idk if they are the finals). Sorted some sprites.

Fixed an interactable bug where you couldn't interact with them after moving.

Fixed a rainbow bug where the particles where not showing up

Changed cam follow so that it returns to the most recent ally instead of the first ally