Tried to fix a few bugs and tweaked some difficulty things
Bug 1) So rooms would go black even when there was no one left in them. I thought it might be because characters were not properly removed from the room when they died and thus titled them "ghost rooms." It turns out it was because the Room script would only remove enemies from the room if the room was active, but would add them whether it was active or not. So enemies (healers) would run into an inactive room and would then run out of the room and would never be removed from the list. The reason we tested if the room was active was because enemies are toggled off at the beginning of the game and that toggle would remove them from the list if we did not have the test. I changed it to test if they were active in hierarchy (so dying or just walked out of the room). I think this fixed it.

Bug 2) When going between floors, we did not retain potions we got from campfires. I looked into it and there were two variables that represented the number of potions. A private one in the main potions script and a public on in a script that was attached to a don't destroy on load object. I removed the private one and now have the main potions script only reference that public one (not the best way, but it was easy and we are low on time).

Bug 3) This isn't really a bug. When you reached max speed and tried to increase speed more, it would change the number text with "7 MAX". This just didn't fit, so I changed it to "MAX".

Bug 4) If a character is selected and then you level up their speed stat, the change is not instantly reflected. To fix this I added an event to the CharDetailMenuController and have the MoveAttackGUIController reference it so that when the event happens, you deselect and reselect the character (if there is one).

Bug 5) Push skill preview displays damage incorrectly. I changed it to set its damage after generation now instead of on awake.

Tweak 1) Changed magic to only have 1 bubble instead of 2. I cleared the game without using any skills (and it was easier than using the skills) and thought it was more boring that way, so I thought this would encourage players to buff their cool skills more.

Tweak 2) The floor before the boss has enemies with like 8-10 health that do 6 damage each. The boss in comparison is not that big a deal. So I changed the boss a bit.